[DS-2701] oai Rip out DB-based search, move general classes out of un-needed .database packages.
