+++ dspace-oai/src/main/java/org/dspace/xoai/app/BasicConfiguration.java
@@ -7,6 +7,14 @@
  */
 package org.dspace.xoai.app;
 
+import org.dspace.xoai.services.impl.DSpaceFieldResolver;
+import org.dspace.xoai.services.impl.DSpaceHandleResolver;
+import org.dspace.xoai.services.impl.DSpaceEarliestDateResolver;
+import org.dspace.xoai.services.impl.DSpaceCollectionsService;
+import org.dspace.xoai.services.api.EarliestDateResolver;
+import org.dspace.xoai.services.api.CollectionsService;
+import org.dspace.xoai.services.api.HandleResolver;
+import org.dspace.xoai.services.api.FieldResolver;
 import com.lyncode.xoai.dataprovider.services.api.ResourceResolver;
 import org.apache.log4j.Logger;
 import org.dspace.xoai.services.api.cache.XOAICacheService;
@@ -16,7 +24,6 @@ import org.dspace.xoai.services.api.config.ConfigurationService;
 import org.dspace.xoai.services.api.config.XOAIManagerResolver;
 import org.dspace.xoai.services.api.config.XOAIManagerResolverException;
 import org.dspace.xoai.services.api.context.ContextService;
-import org.dspace.xoai.services.api.database.*;
 import org.dspace.xoai.services.api.solr.SolrQueryResolver;
 import org.dspace.xoai.services.api.solr.SolrServerResolver;
 import org.dspace.xoai.services.api.xoai.DSpaceFilterResolver;
@@ -30,7 +37,6 @@ import org.dspace.xoai.services.impl.cache.DSpaceXOAILastCompilationCacheService
 import org.dspace.xoai.services.impl.config.DSpaceConfigurationService;
 import org.dspace.xoai.services.impl.context.DSpaceContextService;
 import org.dspace.xoai.services.impl.context.DSpaceXOAIManagerResolver;
-import org.dspace.xoai.services.impl.database.*;
 import org.dspace.xoai.services.impl.resources.DSpaceResourceResolver;
 import org.dspace.xoai.services.impl.solr.DSpaceSolrQueryResolver;
 import org.dspace.xoai.services.impl.solr.DSpaceSolrServerResolver;
@@ -126,7 +132,7 @@ public class BasicConfiguration {
 
     @Bean
     public HandleResolver handleResolver () {
-        return new DSpaceHandlerResolver();
+        return new DSpaceHandleResolver();
     }
 
     @Bean
@@ -137,9 +143,5 @@ public class BasicConfiguration {
     @Bean
     public SolrQueryResolver solrQueryResolver () {
         return new DSpaceSolrQueryResolver();
-    }
-    @Bean
-    public DatabaseQueryResolver databaseQueryResolver () {
-        return new DSpaceDatabaseQueryResolver();
     }
 }
