+++ dspace-oai/src/main/java/org/dspace/xoai/filter/DSpaceFilter.java
@@ -15,7 +15,7 @@ import org.dspace.core.Context;
 import org.dspace.xoai.data.DSpaceItem;
 import org.dspace.xoai.filter.results.DatabaseFilterResult;
 import org.dspace.xoai.filter.results.SolrFilterResult;
-import org.dspace.xoai.services.api.database.FieldResolver;
+import org.dspace.xoai.services.api.FieldResolver;
 
 /**
  *
