+++ dspace-oai/src/main/java/org/dspace/xoai/services/api/HandleResolver.java
@@ -5,7 +5,7 @@
  *
  * http://www.dspace.org/license/
  */
-package org.dspace.xoai.services.api.database;
+package org.dspace.xoai.services.api;
 
 import org.dspace.content.DSpaceObject;
