--- dspace-oai/src/main/java/org/dspace/xoai/services/api/database/DatabaseQueryException.java
@@ -1,25 +1,0 @@
-/**
- * The contents of this file are subject to the license and copyright
- * detailed in the LICENSE and NOTICE files at the root of the source
- * tree and available online at
- *
- * http://www.dspace.org/license/
- */
-package org.dspace.xoai.services.api.database;
-
-public class DatabaseQueryException extends Exception {
-    public DatabaseQueryException() {
-    }
-
-    public DatabaseQueryException(String message) {
-        super(message);
-    }
-
-    public DatabaseQueryException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DatabaseQueryException(Throwable cause) {
-        super(cause);
-    }
-}
