+++ dspace-oai/src/main/java/org/dspace/xoai/services/impl/MetadataFieldCache.java
@@ -5,7 +5,7 @@
  *
  * http://www.dspace.org/license/
  */
-package org.dspace.xoai.services.impl.database;
+package org.dspace.xoai.services.impl;
 
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
