+++ dspace-oai/src/main/java/org/dspace/xoai/services/impl/xoai/DSpaceItemRepositoryResolver.java
@@ -9,13 +9,11 @@ package org.dspace.xoai.services.impl.xoai;
 
 import com.lyncode.xoai.dataprovider.services.api.ItemRepository;
 import org.apache.solr.client.solrj.SolrServerException;
-import org.dspace.xoai.services.api.cache.XOAIItemCacheService;
 import org.dspace.xoai.services.api.config.ConfigurationService;
 import org.dspace.xoai.services.api.context.ContextService;
 import org.dspace.xoai.services.api.context.ContextServiceException;
-import org.dspace.xoai.services.api.database.CollectionsService;
-import org.dspace.xoai.services.api.database.DatabaseQueryResolver;
-import org.dspace.xoai.services.api.database.HandleResolver;
+import org.dspace.xoai.services.api.CollectionsService;
+import org.dspace.xoai.services.api.HandleResolver;
 import org.dspace.xoai.services.api.solr.SolrQueryResolver;
 import org.dspace.xoai.services.api.solr.SolrServerResolver;
 import org.dspace.xoai.services.api.xoai.ItemRepositoryResolver;
@@ -31,13 +29,9 @@ public class DSpaceItemRepositoryResolver implements ItemRepositoryResolver {
     @Autowired
     SolrQueryResolver solrQueryResolver;
     @Autowired
-    DatabaseQueryResolver databaseQueryResolver;
-    @Autowired
     CollectionsService collectionsService;
     @Autowired
     private HandleResolver handleResolver;
-    @Autowired
-    private XOAIItemCacheService cacheService;
 
     private ItemRepository itemRepository;
 
@@ -45,15 +39,15 @@ public class DSpaceItemRepositoryResolver implements ItemRepositoryResolver {
     @Override
     public ItemRepository getItemRepository() throws ContextServiceException {
         if (itemRepository == null) {
-            String storage = configurationService.getProperty("oai", "storage");
-            if (storage == null || !storage.trim().toLowerCase().equals("database")) {
-                try {
-                    itemRepository = new DSpaceItemSolrRepository(solrServerResolver.getServer(), collectionsService, handleResolver, solrQueryResolver);
-                } catch (SolrServerException e) {
-                    throw new ContextServiceException(e.getMessage(), e);
-                }
-            } else
-                itemRepository = new DSpaceItemDatabaseRepository(configurationService, collectionsService, handleResolver, cacheService, databaseQueryResolver, contextService);
+            try {
+                itemRepository = new DSpaceItemSolrRepository(
+                        solrServerResolver.getServer(),
+                        collectionsService,
+                        handleResolver,
+                        solrQueryResolver);
+            } catch (SolrServerException e) {
+                throw new ContextServiceException(e.getMessage(), e);
+            }
         }
 
         return itemRepository;
