+++ dspace-oai/src/test/java/org/dspace/xoai/tests/helpers/stubs/StubbedFieldResolver.java
@@ -9,7 +9,7 @@ package org.dspace.xoai.tests.helpers.stubs;
 
 import org.dspace.core.Context;
 import org.dspace.xoai.exceptions.InvalidMetadataFieldException;
-import org.dspace.xoai.services.api.database.FieldResolver;
+import org.dspace.xoai.services.api.FieldResolver;
 
 import java.sql.SQLException;
 import java.util.HashMap;
