+++ dspace-oai/src/test/java/org/dspace/xoai/tests/integration/xoai/AbstractDSpaceTest.java
@@ -16,8 +16,8 @@ import org.apache.solr.client.solrj.SolrServer;
 import org.dspace.xoai.controller.DSpaceOAIDataProvider;
 import org.dspace.xoai.services.api.config.ConfigurationService;
 import org.dspace.xoai.services.api.config.XOAIManagerResolver;
-import org.dspace.xoai.services.api.database.EarliestDateResolver;
-import org.dspace.xoai.services.api.database.FieldResolver;
+import org.dspace.xoai.services.api.EarliestDateResolver;
+import org.dspace.xoai.services.api.FieldResolver;
 import org.dspace.xoai.tests.DSpaceTestConfiguration;
 import org.dspace.xoai.tests.helpers.stubs.*;
 import org.junit.After;
