+++ dspace-oai/src/test/java/org/dspace/xoai/tests/unit/services/impl/AbstractQueryResolverTest.java
@@ -10,9 +10,9 @@ package org.dspace.xoai.tests.unit.services.impl;
 import com.lyncode.builder.DateBuilder;
 import com.lyncode.xoai.dataprovider.services.impl.BaseDateProvider;
 import org.apache.solr.client.solrj.util.ClientUtils;
-import org.dspace.xoai.services.api.database.CollectionsService;
-import org.dspace.xoai.services.api.database.FieldResolver;
-import org.dspace.xoai.services.api.database.HandleResolver;
+import org.dspace.xoai.services.api.CollectionsService;
+import org.dspace.xoai.services.api.FieldResolver;
+import org.dspace.xoai.services.api.HandleResolver;
 import org.dspace.xoai.services.api.xoai.DSpaceFilterResolver;
 import org.dspace.xoai.tests.DSpaceBasicTestConfiguration;
 import org.dspace.xoai.tests.helpers.stubs.StubbedFieldResolver;
